% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_tbl_has_columns}
\alias{assert_tbl_has_columns}
\title{Assert Tibble Has Columns}
\usage{
assert_tbl_has_columns(tbl, columns)
}
\arguments{
\item{tbl}{A tibble}

\item{columns}{a vector of columns}
}
\description{
Assert Tibble Has Columns
}
