% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_outcomes_functions.R
\name{build_co_heatmap_matrix}
\alias{build_co_heatmap_matrix}
\title{Build Heatmap Matrix}
\usage{
build_co_heatmap_matrix(tbl)
}
\arguments{
\item{tbl}{A tibble with columns feature, value, time, status, group}
}
\description{
Build Heatmap Matrix
}
