% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_outcomes_functions.R
\name{build_co_survival_tbl}
\alias{build_co_survival_tbl}
\title{Build Clinical Outcomes Survival Tibble}
\usage{
build_co_survival_tbl(time_id, status_id)
}
\arguments{
\item{time_id}{An integer in the id column of the samples table}

\item{status_id}{An integer in the id column of the samples table}
}
\description{
Build Clinical Outcomes Survival Tibble
}
