% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_outcomes_functions.R
\name{build_co_survival_value_tbl}
\alias{build_co_survival_value_tbl}
\title{Build Survival Values Tibble}
\usage{
build_co_survival_value_tbl(cohort_obj, time, status)
}
\arguments{
\item{sample_tbl}{A Tibble with columns sample_id and group}

\item{time_id}{An integer in the id column of the samples table}

\item{status_id}{An integer in the id column of the samples table}
}
\description{
Build Survival Values Tibble
}
