% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_cohort_object}
\alias{build_cohort_object}
\title{Build Cohort Object}
\usage{
build_cohort_object(filter_obj, dataset, group_choice, group_type, ...)
}
\arguments{
\item{filter_obj}{A named list with element sample_ids and filters}

\item{dataset}{A string}

\item{group_choice}{A string}

\item{driver_mutation}{A string}

\item{immune_feature_bin_id}{An integer}

\item{immune_feature_bin_number}{An integer}
}
\description{
Build Cohort Object
}
