% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_cohort_tbl_by_tag}
\alias{build_cohort_tbl_by_tag}
\title{Build Cohort Tibble By Tag}
\usage{
build_cohort_tbl_by_tag(samples, dataset, tag)
}
\arguments{
\item{sample_ids}{Integers in the id column of the samples table}

\item{group}{A String that is the display column of the tags table}
}
\description{
Build Cohort Tibble By Tag
}
