% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/tumor_microenvironment_type_fractions_functions.R
\name{build_ctf_barplot_tbl}
\alias{build_ctf_barplot_tbl}
\title{Build Cell Type Fractions Barplot Tibble}
\usage{
build_ctf_barplot_tbl(tbl)
}
\arguments{
\item{class_name}{A string, that is the name column of the classes table}

\item{sample_tbl}{A tibble with columns sample_id, and group}
}
\description{
Build Cell Type Fractions Barplot Tibble
}
