% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_dm_cohort_object}
\alias{build_dm_cohort_object}
\title{Build Driver Mutation Cohort Object}
\usage{
build_dm_cohort_object(dataset, samples, mutation_id, mutation_tbl)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{mutation}{A string}
}
\description{
Build Driver Mutation Cohort Object
}
