% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_functions.R
\name{build_driver_violin_tbl}
\alias{build_driver_violin_tbl}
\title{Build Driver Violin Tibble}
\usage{
build_driver_violin_tbl(feature_id, gene_id, tag_id, mutation_id)
}
\arguments{
\item{feature_id}{An integer in the features_to_samples table}

\item{gene_id}{An interger in the genes_samples_mutations table}

\item{tag_id}{An integer in the samples_to_tags table}

\item{mutation_id}{An integer in the genes_samples_mutations table}
}
\description{
Build Driver Violin Tibble
}
