% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_manual_selection_functions.R
\name{build_feature_bin_group_tbl}
\alias{build_feature_bin_group_tbl}
\title{Build Feature Bin Group Tibble}
\usage{
build_feature_bin_group_tbl(sample_tbl, feature_name)
}
\arguments{
\item{sample_tbl}{A tibble with name group}

\item{feature_name}{A string}
}
\description{
Build Feature Bin Group Tibble
}
