% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{build_feature_tbl}
\alias{build_feature_tbl}
\title{Build Feature Tibble}
\usage{
build_feature_tbl(class_ids = "all")
}
\arguments{
\item{class_ids}{Integers in the id column of the classes table}
}
\description{
Build Feature Tibble
}
