% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_functions.R
\name{build_group_filter_tbl}
\alias{build_group_filter_tbl}
\title{Build Group Filter Tibble}
\usage{
build_group_filter_tbl(feature_id)
}
\arguments{
\item{feature_id}{An integer in the related_tag_id column of the
tags_to_tags table}
}
\description{
Build Group Filter Tibble
}
