% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_correlations_functions.R
\name{build_ifc_scatterplot_tbl}
\alias{build_ifc_scatterplot_tbl}
\title{Build Immune Feature Scatterplot Tibble}
\usage{
build_ifc_scatterplot_tbl(tbl, .feature, .group)
}
\arguments{
\item{tbl}{A tibble with columns feature_display,
sample, tag, response_value, feature_value}

\item{.feature}{A string that is in the tbl feature_name column}

\item{.group}{A string that is in the tbl group column}
}
\description{
Build Immune Feature Scatterplot Tibble
}
