% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_distributions_functions.R
\name{build_ifd_distplot_tbl}
\alias{build_ifd_distplot_tbl}
\title{Build Immune Feature Distribution Plot Tibble}
\usage{
build_ifd_distplot_tbl(sample_tbl, feature_name, scale_method)
}
\arguments{
\item{sample_tbl}{A tibble with columns "sample_name", "group"}

\item{feature_name}{A name in the features table}

\item{scale_method}{A method for scaling the value column}
}
\description{
Build Immune Feature Distribution Plot Tibble
}
