% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_target_functions.R
\name{build_io_target_dt_tbl}
\alias{build_io_target_dt_tbl}
\title{Build IO Target Datatable Tibble}
\usage{
build_io_target_dt_tbl(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Build IO Target Datatable Tibble
}
