% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{build_md_covariate_tbl}
\alias{build_md_covariate_tbl}
\title{Build Multivariate Driver Covariate Tibble}
\usage{
build_md_covariate_tbl(cohort_obj, cov_obj)
}
\arguments{
\item{cov_obj}{A list with items named categorical_covariates and
numerical_covariates}
}
\description{
Build Multivariate Driver Covariate Tibble
}
