% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{build_md_feature_covariate_tbl}
\alias{build_md_feature_covariate_tbl}
\title{Build Multivariate Driver Feature Covariate Tibble}
\usage{
build_md_feature_covariate_tbl(cohort_obj, cov_obj)
}
\arguments{
\item{covariates}{A vector of integers that are in the id column of the
features table}
}
\description{
Build Multivariate Driver Feature Covariate Tibble
}
