% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{build_md_pvalue_tbl}
\alias{build_md_pvalue_tbl}
\title{Build Multivariate Driver Pvalue Tibble}
\usage{
build_md_pvalue_tbl(tbl, formula_string)
}
\arguments{
\item{tbl}{A tibble}

\item{formula_string}{A string that could be converted to a formula}
}
\description{
Build Multivariate Driver Pvalue Tibble
}
