% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{build_md_response_tbl}
\alias{build_md_response_tbl}
\title{Build Multivariate Driver Response Tibble}
\usage{
build_md_response_tbl(cohort_obj, feature)
}
\arguments{
\item{feature_id}{An integer in the feature_id column of the
features_to_samples table}
}
\description{
Build Multivariate Driver Response Tibble
}
