% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_functions.R
\name{build_md_results_tbl}
\alias{build_md_results_tbl}
\title{Build Multivariate Driver Results Tibble}
\usage{
build_md_results_tbl(filtered_tbl, pvalue_tbl, effect_size_tbl)
}
\arguments{
\item{filtered_tbl}{A tibble}

\item{pvalue_tbl}{A tibble}

\item{effect_size_tbl}{A tibble}
}
\description{
Build Multivariate Driver Results Tibble
}
