% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_tag_cohort_object}
\alias{build_tag_cohort_object}
\title{Build Tag Cohort Object}
\usage{
build_tag_cohort_object(dataset, samples, tag)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{group_choice}{A string}
}
\description{
Build Tag Cohort Object
}
