% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_functions.R
\name{build_tag_filter_list}
\alias{build_tag_filter_list}
\title{Build Tag Filter Named List}
\usage{
build_tag_filter_list(parent_tag_name, dataset)
}
\arguments{
\item{parent_tag_name}{A string}
}
\description{
Build Tag Filter Named List
}
