% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/til_map_distributions_functions.R
\name{build_tm_distplot_tbl}
\alias{build_tm_distplot_tbl}
\title{Build Tilmap Distplot Tibble}
\usage{
build_tm_distplot_tbl(tbl, id, scale_method)
}
\arguments{
\item{tbl}{A tibble with columns sample_id, sample_name, slide_barcode, group}

\item{id}{An integer in the feature_id column of the features_to_samples
table}

\item{scale_method}{A string}
}
\description{
Build Tilmap Distplot Tibble
}
