% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_driver_functions.R
\name{build_ud_results_tbl}
\alias{build_ud_results_tbl}
\title{Build Univariate Driver Results Tibble}
\usage{
build_ud_results_tbl(group_name, feature_id, min_wt, min_mut)
}
\arguments{
\item{group_name}{A string in the name column of the tags table}

\item{feature_id}{An integer in the feature_id column of the driver_results
table}

\item{min_wt}{An integer}

\item{min_mut}{An integer}
}
\description{
Build Univariate Driver Results Tibble
}
