% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_functions.R
\name{calculate_lm_pvalue}
\alias{calculate_lm_pvalue}
\title{Calculate Linear Model Pvalue}
\usage{
calculate_lm_pvalue(data, lm_formula, term)
}
\arguments{
\item{data}{A dataframe}

\item{lm_formula}{A string or formula}

\item{term}{A string}
}
\description{
Calculate Linear Model Pvalue
}
