% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunomodulators_functions.R
\name{create_build_im_tbl_query}
\alias{create_build_im_tbl_query}
\title{Create Build Immunomodulators Table Query}
\usage{
create_build_im_tbl_query()
}
\description{
Create Build Immunomodulators Table Query
}
