% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{create_class_list}
\alias{create_class_list}
\title{Build Gene Tibble}
\usage{
create_class_list()
}
\arguments{
\item{sample_ids}{Integers in the sample_id column of the
genes_to_samples table
Create Class List}
}
\description{
Build Gene Tibble
}
