% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{create_cohort_filter_object}
\alias{create_cohort_filter_object}
\title{Create Cohort Filter Object}
\usage{
create_cohort_filter_object(sample_ids, numeric_obj, group_obj)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{numeric_obj}{A list}

\item{group_obj}{A list}
}
\description{
Create Cohort Filter Object
}
