% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{create_cohort_group_named_list}
\alias{create_cohort_group_named_list}
\title{Create Cohorts Group Named List}
\usage{
create_cohort_group_named_list(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Create Cohorts Group Named List
}
