% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_dataset_selection_functions.R
\name{create_cohort_module_string}
\alias{create_cohort_module_string}
\title{Create Cohort Module String}
\usage{
create_cohort_module_string(tbl, .dataset)
}
\arguments{
\item{tbl}{A tibble with columns dataset, and module}

\item{.dataset}{a string in the dataset column of the tibble}
}
\description{
Create Cohort Module String
}
