% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_cohort_object}
\alias{create_cohort_object}
\title{Create Cohort Object}
\usage{
create_cohort_object(
  filter_obj,
  group_choice,
  dataset,
  driver_mutation = NULL,
  immune_feature_bin_id = NULL,
  immune_feature_bin_number = NULL
)
}
\arguments{
\item{filter_obj}{A named list with element sample_ids and filters}

\item{group_choice}{A string}

\item{dataset}{A string}

\item{driver_mutation}{A string}

\item{immune_feature_bin_id}{An integer}

\item{immune_feature_bin_number}{An integer}
}
\description{
Create Cohort Object
}
