% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_correlated_subquery}
\alias{create_correlated_subquery}
\title{Create Correlated Subquery}
\usage{
create_correlated_subquery(table, into, from, value, new_column)
}
\description{
Create Correlated Subquery
}
