% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{create_covariate_string}
\alias{create_covariate_string}
\title{Create Categorical Covariate String}
\usage{
create_covariate_string(prefix_str, num_str, cat_str)
}
\arguments{
\item{prefix_str}{A string}

\item{num_str}{A string or NULL}

\item{cat_str}{A string or NULL}
}
\description{
Create Categorical Covariate String
}
