% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_dm_cohort_group_tbl}
\alias{create_dm_cohort_group_tbl}
\title{Create Driver Mutation Cohort Group Tibbl}
\usage{
create_dm_cohort_group_tbl(sample_tbl, gene_name)
}
\arguments{
\item{sample_tbl}{A tibble}

\item{gene_name}{A string}
}
\description{
Create Driver Mutation Cohort Group Tibbl
}
