% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_dm_cohort_sample_tbl}
\alias{create_dm_cohort_sample_tbl}
\title{Create Driver Mutation Cohort Sample Tibble}
\usage{
create_dm_cohort_sample_tbl(sample_ids, gene_id, code_id)
}
\arguments{
\item{sample_ids}{A vecotr of itnegers}

\item{gene_id}{An integer}

\item{code_id}{An integer}
}
\description{
Create Driver Mutation Cohort Sample Tibble
}
