% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_feature_bin_sample_tbl}
\alias{create_feature_bin_sample_tbl}
\title{Create Feature Bin Sample Tibble}
\usage{
create_feature_bin_sample_tbl(sample_ids, feature_id, n_bins)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{feature_id}{An integer}

\item{n_bins}{An integer}
}
\description{
Create Feature Bin Sample Tibble
}
