% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_get_feature_display_from_id_query}
\alias{create_get_feature_display_from_id_query}
\title{Create Get Feature Display from Id Query}
\usage{
create_get_feature_display_from_id_query(id)
}
\arguments{
\item{id}{A feature id}
}
\description{
Create Get Feature Display from Id Query
}
