% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{create_group_text_from_eventdata}
\alias{create_group_text_from_eventdata}
\title{Create Group Text from Eventdata Dataframe}
\usage{
create_group_text_from_eventdata(eventdata, group_tbl)
}
\arguments{
\item{eventdata}{Eventdata from "plotly_click" plotly::event_data}

\item{group_tbl}{A Tibble with columns group, name, characteristics}
}
\description{
Create Group Text from Eventdata Dataframe
}
