% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_target_functions.R
\name{create_io_landscape_links}
\alias{create_io_landscape_links}
\title{Create IO Landscape Links}
\usage{
create_io_landscape_links(io_gene_names)
}
\arguments{
\item{genes}{A vector of strings
paste0("<a href='",mydata$url,"'>",mydata$url,"</a>")}
}
\description{
Create IO Landscape Links
}
