% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_functions.R
\name{create_md_violin_plot_title}
\alias{create_md_violin_plot_title}
\title{Create Multivariate Driver Violin Plot Title}
\usage{
create_md_violin_plot_title(tbl, mode)
}
\arguments{
\item{tbl}{A one row tibble with columns p_value and log10_fold_change
if mode == "By group" The tibble must has a group column}

\item{mode}{A string, either "By group" or "Across groups"}
}
\description{
Create Multivariate Driver Violin Plot Title
}
