% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{create_md_violin_plot_x_lab}
\alias{create_md_violin_plot_x_lab}
\title{Create Multivariate Driver Violin Plot X Label}
\usage{
create_md_violin_plot_x_lab(label, mode)
}
\arguments{
\item{label}{A string}

\item{mode}{A string, either "By group" or "Across groups"}
}
\description{
Create Multivariate Driver Violin Plot X Label
}
