% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_parent_group_query_from_id}
\alias{create_parent_group_query_from_id}
\title{Create Parent Group Query From Id}
\usage{
create_parent_group_query_from_id(id)
}
\arguments{
\item{id}{The id of the parent group}
}
\description{
Create Parent Group Query From Id
}
