% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_plot_colors_list}
\alias{create_plot_colors_list}
\title{Create Plot Colors List}
\usage{
create_plot_colors_list(tbl)
}
\arguments{
\item{tbl}{A tibble with name group, and optionally color}
}
\description{
Create Plot Colors List
}
