% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_tag_group_tbl}
\alias{create_tag_group_tbl}
\title{Create Tag Group Tibble}
\usage{
create_tag_group_tbl(cohort_tbl)
}
\arguments{
\item{cohort_tbl}{A Tibble with columns group, name, characteristics}
}
\description{
Create Tag Group Tibble
}
