% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/til_map_distributions_functions.R
\name{create_tm_named_list}
\alias{create_tm_named_list}
\title{Create Tilmap Named List}
\usage{
create_tm_named_list(tbl)
}
\arguments{
\item{tbl}{A tibble with columns class, display and id}
}
\description{
Create Tilmap Named List
}
