% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_translate_values_query}
\alias{create_translate_values_query}
\title{Create Translate Values Query}
\usage{
create_translate_values_query(table, into, from, query)
}
\arguments{
\item{table}{The name of the table in the database to query}

\item{into}{The column in the table to translate into}

\item{from}{The column in the table to translate from}

\item{query}{A string that is a valid sql query that results in a one column
table that contains the from value}
}
\description{
Create Translate Values Query
}
