% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_functions.R
\name{format_filtered_feature_tbl}
\alias{format_filtered_feature_tbl}
\title{Format Filtered Feature Tibble}
\usage{
format_filtered_feature_tbl(filtered_feature_tbl)
}
\arguments{
\item{filtered_feature_tbl}{A tibble with columns order, feature, unit, and
method_tag}
}
\description{
Format Filtered Feature Tibble
}
