% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{get_code_from_mutation}
\alias{get_code_from_mutation}
\title{Get Code From Mutation}
\usage{
get_code_from_mutation(mutation)
}
\arguments{
\item{mutation}{A string}
}
\description{
Get Code From Mutation
}
