% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{get_cohort_available_groups}
\alias{get_cohort_available_groups}
\title{Get Cohort Available Groups}
\usage{
get_cohort_available_groups(tbl, .dataset)
}
\arguments{
\item{tbl}{A tibble}

\item{.dataset}{A string}
}
\description{
Get Cohort Available Groups
}
