% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{get_effect_size_from_tbl}
\alias{get_effect_size_from_tbl}
\title{Get Effect Size From Tibble}
\usage{
get_effect_size_from_tbl(tbl, method = calculate_ratio_effect_size)
}
\arguments{
\item{tbl}{A tibble}

\item{method}{A function}
}
\description{
Get Effect Size From Tibble
}
