% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_feature_display_from_id}
\alias{get_feature_display_from_id}
\title{Get Feature Display Name From ID}
\usage{
get_feature_display_from_id(id)
}
\arguments{
\item{id}{An integer in the id column of the features table}
}
\description{
Get Feature Display Name From ID
}
