% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_filtered_feature_sample_ids}
\alias{get_filtered_feature_sample_ids}
\title{Get Filtered Feature Sample IDs}
\usage{
get_filtered_feature_sample_ids(filter_obj, sample_ids)
}
\arguments{
\item{filter_obj}{A list of named lists with names id max and min}

\item{sample_ids}{A vector of integers in the sample_id column of the
features_to_samples table}
}
\description{
Get Filtered Feature Sample IDs
}
