% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_filtered_tag_samples}
\alias{get_filtered_tag_samples}
\title{Get Filtered Tag Samples}
\usage{
get_filtered_tag_samples(filter_obj, samples, dataset)
}
\arguments{
\item{filter_obj}{A list of named lists with names ids and name}

\item{samples}{A vector of strings}
}
\description{
Get Filtered Tag Samples
}
