% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_target_functions.R
\name{get_gene_from_url}
\alias{get_gene_from_url}
\title{Get Gene From URL}
\usage{
get_gene_from_url(url_query)
}
\arguments{
\item{url_query}{A list with named value gene}
}
\description{
Get Gene From URL
}
