% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_gene_hgnc_from_id}
\alias{get_gene_hgnc_from_id}
\title{Get HGNC Symbol from Gene ID}
\usage{
get_gene_hgnc_from_id(gene_id)
}
\arguments{
\item{gene_id}{A integer from the id column of the genes table}
}
\description{
Get HGNC Symbol from Gene ID
}
