% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_gene_id_from_hgnc}
\alias{get_gene_id_from_hgnc}
\title{Get Gene ID from HGNC Symbol}
\usage{
get_gene_id_from_hgnc(hgnc)
}
\arguments{
\item{hgnc}{A string in the hgnc column of the genes table}
}
\description{
Get Gene ID from HGNC Symbol
}
