% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_mutation_code_from_id}
\alias{get_mutation_code_from_id}
\title{Get Mutation Code From ID}
\usage{
get_mutation_code_from_id(id)
}
\arguments{
\item{id}{An integer in the id column of the mutation_codes table}
}
\description{
Get Mutation Code From ID
}
