% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_valid_tag_filters}
\alias{get_valid_tag_filters}
\title{Get Valid Tag Filters}
\usage{
get_valid_tag_filters(filter_obj)
}
\arguments{
\item{filter_obj}{A list of named lists with names ids and name}
}
\description{
Get Valid Tag Filters
}
