% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{is_group_filter_valid}
\alias{is_group_filter_valid}
\title{Is Group Filter Valid}
\usage{
is_group_filter_valid(obj)
}
\arguments{
\item{obj}{A named list with names ids and name}
}
\description{
Is Group Filter Valid
}
