% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{is_numeric_filter_valid}
\alias{is_numeric_filter_valid}
\title{Is Numeric Filter Valid}
\usage{
is_numeric_filter_valid(obj)
}
\arguments{
\item{obj}{A named lists with names id max and min}
}
\description{
Is Numeric Filter Valid
}
