% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{is_tag_filter_valid}
\alias{is_tag_filter_valid}
\title{Is Tag Filter Valid}
\usage{
is_tag_filter_valid(obj)
}
\arguments{
\item{obj}{A named list with names}
}
\description{
Is Tag Filter Valid
}
