% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{numeric_values_to_query_list}
\alias{numeric_values_to_query_list}
\title{Numeric Values to Query List}
\usage{
numeric_values_to_query_list(values)
}
\arguments{
\item{values}{A numeric vector}
}
\description{
Numeric Values to Query List
}
