% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_query.R
\name{perform_query}
\alias{perform_query}
\title{Perform Query}
\usage{
perform_query(query, message = "", db_pool = .GlobalEnv$pool)
}
\arguments{
\item{query}{A string that is a valid query}

\item{message}{A string that is displayed when run}

\item{db_pool}{A database ppol of type pool::dbPool}
}
\description{
Perform Query
}
