% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_tbl_value_column}
\alias{scale_tbl_value_column}
\title{Scale Tibble Value Column}
\usage{
scale_tbl_value_column(tbl, scale_method = "None")
}
\arguments{
\item{tbl}{A Tibble with column "value"}

\item{scale_method}{One of "Log2", "Log2 + 1", "Log10 + 1", "Log10"}
}
\description{
Scale Tibble Value Column
}
