% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{string_values_to_query_list}
\alias{string_values_to_query_list}
\title{String Values to Query List}
\usage{
string_values_to_query_list(values)
}
\arguments{
\item{values}{A character vector}
}
\description{
String Values to Query List
}
