% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_feature_formula}
\alias{transform_feature_formula}
\title{Transform Feature Formula}
\usage{
transform_feature_formula(feature, transformation)
}
\arguments{
\item{feature}{A string, the name of the feature}

\item{transformation}{A string, the name of the transformation}
}
\description{
Transform Feature Formula
}
