% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{translate_value}
\alias{translate_value}
\title{>>>>>>> staging_cohort_selection_ux
#' Get Sample IDs from Dataset
#'
#' @param dataset The name of a dataset in the database
#' @importFrom magrittr %>%
#' @importFrom dplyr pull
#' @importFrom rlang .data
get_sample_ids_from_dataset <- function(dataset){
    paste0(
<<<<<<< HEAD
        "SELECT sample_id FROM samples_to_tags WHERE tag_id IN ",
        "(SELECT id FROM tags where name = '", dataset, "')"
    ) %>%
        perform_query("Get Sample IDs from Dataset") %>%
=======
        "SELECT * FROM samples_to_tags WHERE tag_id IN (",
        "SELECT id FROM tags WHERE display IN ('",
        dataset,
        "')",
        ")"
    ) %>%
        perform_query() %>%
>>>>>>> staging_cohort_selection_ux
        dplyr::pull(.data$sample_id)
}}
\usage{
translate_value(value, value_type, table, into, from)
}
\description{
>>>>>>> staging_cohort_selection_ux
#' Get Sample IDs from Dataset
#'
#' @param dataset The name of a dataset in the database
#' @importFrom magrittr %>%
#' @importFrom dplyr pull
#' @importFrom rlang .data
get_sample_ids_from_dataset <- function(dataset){
    paste0(
<<<<<<< HEAD
        "SELECT sample_id FROM samples_to_tags WHERE tag_id IN ",
        "(SELECT id FROM tags where name = '", dataset, "')"
    ) %>%
        perform_query("Get Sample IDs from Dataset") %>%
=======
        "SELECT * FROM samples_to_tags WHERE tag_id IN (",
        "SELECT id FROM tags WHERE display IN ('",
        dataset,
        "')",
        ")"
    ) %>%
        perform_query() %>%
>>>>>>> staging_cohort_selection_ux
        dplyr::pull(.data$sample_id)
}
}
