% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{create_numerical_covariate_string}
\alias{create_numerical_covariate_string}
\title{Create Numerical Covariate String}
\usage{
create_numerical_covariate_string(covs, transforms, transform_func)
}
\arguments{
\item{covs}{A vector of strings or NULL}

\item{transforms}{A vector of strings or NULL}

\item{transform_func}{A function}

\item{translate_func}{A function}
}
\description{
Create Numerical Covariate String
}
