% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_filtered_feature_sample_ids}
\alias{get_filtered_feature_sample_ids}
\title{Get Filtered Feature Sample IDs}
\usage{
get_filtered_feature_sample_ids(feature_id, min, max)
}
\arguments{
\item{feature_id}{An integer in the smaple_id column of the
features_to_samples table}

\item{min}{A numeric}

\item{max}{A numeric}
}
\description{
Get Filtered Feature Sample IDs
}
